修复数据库中字段值为空时无法搜索到数据的问题

+++ academy-canary-core/src/main/java/com/ptteng/academy/business/dto/StudyDto.java
@@ -19,7 +19,7 @@ public class StudyDto {
     private String title; // 标题
     private String cover_plan_url; // 封面
     private Integer classify; // 分类
-    private Integer author;  // 作者
+    private Long author;  // 作者
     private String author_img; // 作者头像
     private String author_name; // 作者名称
     private String introduce; // 摘要
